--- ast/ansi/src/test/scala/scoobie/ProjectionTests.scala
@@ -1,13 +1,0 @@
-package scoobie
-
-import org.specs2._
-
-/**
-  * Created by jacob.barber on 5/23/16.
-  */
-trait ProjectionTests extends SpecificationLike with ParamTests {
-  lazy val queryProjectOneTest = {
-    projection.alias mustEqual None
-    projection.selection mustEqual fooParam
-  }
-}
