--- doobie/postgres/src/main/scala/scoobie/doobie/postgres.scala
@@ -1,25 +1,0 @@
-package scoobie.doobie
-
-import doobie.imports._
-import scoobie.ast.coercion.Coerce
-import scoobie.interpreters.ansi._
-import scoobie.interpreters.ansi.SqlInterpreter.LiteralQueryString
-
-/**
-  * Created by jbarber on 5/20/16.
-  */
-object postgres extends DoobieSupport {
-
-  implicit val fragmentLifter =
-    new SqlQueryLifter[ScoobieFragmentProducer, Fragment] {
-      def liftValue[A](a: A, fa: ScoobieFragmentProducer[A]): Fragment = fa.genFragment(a)
-    }
-
-  implicit val scoobieFragmentProducerForLiteralQueryString: ScoobieFragmentProducer[LiteralQueryString] =
-    ScoobieFragmentProducer[LiteralQueryString](s => new StringContext(s.s).fr0.apply())
-
-  val interpreter: SqlInterpreter[ScoobieFragmentProducer, Fragment] = SqlInterpreter[ScoobieFragmentProducer, Fragment]("\"")
-
-  implicit val doobieInterpreter = DoobieSqlInterpreter(interpreter.interpretSql(_))
-  implicit val coercetoScoobieFragmentProducer: Coerce[ScoobieFragmentProducer] = scoobie.doobie.doobieCoercer
-}
