--- doobie/support/src/it/scala/scoobie/doobie/log.scala
@@ -1,41 +1,0 @@
-package scoobie.doobie
-
-import doobie.util.log._
-import java.util.logging.Logger
-
-object log {
-  val verboseTestLogger: LogHandler = {
-    val logger = Logger.getLogger(getClass.getName)
-    LogHandler {
-
-       case Success(s, a, e1, e2) =>
-          logger.info(s"""Successful Statement Execution:
-            |
-            |  ${s.lines.dropWhile(_.trim.isEmpty).mkString("\n  ")}
-            |
-            | arguments = [${a.mkString(", ")}]
-            |   elapsed = ${e1.toMillis} ms exec + ${e2.toMillis} ms processing (${(e1 + e2).toMillis} ms total)
-          """.stripMargin)
-
-        case ProcessingFailure(s, a, e1, e2, t) =>
-          logger.severe(s"""Failed Resultset Processing:
-            |
-            |  ${s.lines.dropWhile(_.trim.isEmpty).mkString("\n  ")}
-            |
-            | arguments = [${a.mkString(", ")}]
-            |   elapsed = ${e1.toMillis} ms exec + ${e2.toMillis} ms processing (failed) (${(e1 + e2).toMillis} ms total)
-            |   failure = ${t.getMessage}
-          """.stripMargin)
-
-        case ExecFailure(s, a, e1, t) =>
-          logger.severe(s"""Failed Statement Execution:
-            |
-            |  ${s.lines.dropWhile(_.trim.isEmpty).mkString("\n  ")}
-            |
-            | arguments = [${a.mkString(", ")}]
-            |   elapsed = ${e1.toMillis} ms exec (failed)
-            |   failure = ${t.getMessage}
-          """.stripMargin)
-    }
-  }
-}
